Agrego relacion Curso/Materia

+++ backend/src/main/java/com/ips/gestion_academica/dto/materia/MateriaResponse.java
@@ -2,14 +2,14 @@ package com.ips.gestion_academica.dto.materia;
 
 public class MateriaResponse {
 
-    private int id;
+    private long id;
     private String codigo;
     private String nombre;
     private String descripcion;
     private int anioCursada;
     private Boolean activo;
 
-    public MateriaResponse(int id, String codigo, String nombre, String descripcion, int anioCursada, Boolean activo) {
+    public MateriaResponse(long id, String codigo, String nombre, String descripcion, int anioCursada, Boolean activo) {
         this.id = id;
         this.codigo = codigo;
         this.nombre = nombre;
@@ -18,7 +18,7 @@ public class MateriaResponse {
         this.activo = activo;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
