+++ backend/src/main/java/com/ips/gestion_academica/repository/MateriaRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ips.gestion_academica.model.Materia;
 
-public interface MateriaRepository extends JpaRepository<Materia, Integer> {
+public interface MateriaRepository extends JpaRepository<Materia, Long> {
     boolean existsByCodigo(String codigo);
+
 }
